+++ src/MenuCharacter.cpp
@@ -190,7 +190,7 @@ MenuCharacter::MenuCharacter(StatBlock *_stats)
 			else if (infile.key == "show_stat") {
 				std::string stat_name = Parse::popFirstString(infile.val);
 
-				for (unsigned i=0; i<Stats::COUNT; ++i) {
+				for (int i=0; i<Stats::COUNT; ++i) {
 					if (stat_name == Stats::KEY[i]) {
 						show_stat[i] = Parse::toBool(Parse::popFirstString(infile.val));
 						break;
@@ -315,7 +315,7 @@ void MenuCharacter::refreshStats() {
 
 	// scrolling stat list
 	unsigned stat_index = 0;
-	for (unsigned i=0; i<Stats::COUNT; ++i) {
+	for (int i=0; i<Stats::COUNT; ++i) {
 		if (!show_stat[i]) continue;
 
 		// insert damage stats before absorb min
@@ -371,7 +371,7 @@ void MenuCharacter::refreshStats() {
 		cstat[j].tip.addText(cstat[j].label->getText());
 		cstat[j].tip.addText(msg->get("base (%d), bonus (%d)", *(base_stats[j-2]), *(base_stats_add[j-2])));
 		bool have_bonus = false;
-		for (size_t i = 0; i < Stats::COUNT; ++i) {
+		for (int i = 0; i < Stats::COUNT; ++i) {
 			// damage types are displayed before absorb
 			if (i == Stats::ABS_MIN) {
 				for (size_t k = 0; k < eset->damage_types.count; ++k) {
